4.0.0 Removed unsused objects
