+++ src/main/java/com/github/bgora/rpnlibrary/RPNExecuting.java
@@ -23,7 +23,6 @@ import com.github.bgora.rpnlibrary.exceptions.NoSuchFunctionFound;
 import com.github.bgora.rpnlibrary.exceptions.WrongArgumentException;
 
 import java.math.BigDecimal;
-import java.math.RoundingMode;
 
 /**
  * Interface used for executing Calculation on RPN String
@@ -39,21 +38,19 @@ public interface RPNExecuting {
      * @param operator Arithmetic operator to execute.
      * @param var1     first variable
      * @param var2     second variable.
-     * @param mode     Rounding Mode for operation.
      * @return calculation result.
      * @throws WrongArgumentException if theres is something wrong withe the input.
      */
-    BigDecimal executeOperator(String operator, String var1, String var2, RoundingMode mode) throws WrongArgumentException;
+    BigDecimal executeOperator(String operator, String var1, String var2) throws WrongArgumentException;
 
     /**
      * This method executes Arithmetic Functions.
      * The first parameter is Function name.
      *
      * @param functionName Name of the function to execute.
-     * @param mode         Rounding mode for calculations.
      * @param arguments    List of arguments for the called function.
      * @return Calculation result as BigDecimal.
      * @throws NoSuchFunctionFound thrown if Executing object cannot find method.
      */
-    BigDecimal executeFunction(String functionName, RoundingMode mode, String... arguments) throws NoSuchFunctionFound;
+    BigDecimal executeFunction(String functionName, String... arguments) throws NoSuchFunctionFound;
 }
