+++ src/test/java/com/github/bgora/rpnlibrary/AdvancedCalulatorTest2ParamFunction.java
@@ -19,7 +19,6 @@
 
 package com.github.bgora.rpnlibrary;
 
-import com.github.bgora.rpnlibrary.advanced.AdvancedCalculatorFactory;
 import com.github.bgora.rpnlibrary.exceptions.NoSuchFunctionFound;
 import com.github.bgora.rpnlibrary.exceptions.WrongArgumentException;
 import org.junit.Before;
